Misc cleanup (#194)
* Rename `watcher` module to `file_watcher`

The name change itself is largely benign. The sole motivation is that
`cargo test` defaults to alphabetical order for running tests and the
`watcher::tests::the_guantlet` is the longest test by far while also
being the last in the alphabet.

This change makes the test suite go from taking 0.92 seconds to 0.64
seconds. `file_watcher::tests::the_gauntlet` also happened to take 0.64
seconds according to `cargo nextest run`

* Mostly remove bespoke cruft from keybindings sanity test

* Mostly standardized import style

Not actually an important change, but some of the imports were
sprawling, so I figured it was worth doing a little cleanup

* Fix where cache is created on macos arm CI job

* Make yaml snapshot test slightly more succinct

* Also run CI tests on `beta` to catch breakages semi-early

* Placate `clippy`

* Add `clippy` and `rustfmt` checks to CI

* `#![deny]` some useful clippy lints